+++ engine/algebra/abeliangroup.h
@@ -78,7 +78,7 @@ using MatrixInt = Matrix<Integer, true>;
  */
 class AbelianGroup : public ShortOutput<AbelianGroup, true> {
     protected:
-        unsigned rank_ { 0 };
+        unsigned long rank_ { 0 };
             /**< The rank of the group (the number of Z components). */
         std::vector<Integer> revInvFactors_;
             /**< The invariant factors <i>d0</i>,...,<i>dn</i> as
@@ -110,7 +110,7 @@ class AbelianGroup : public ShortOutput<AbelianGroup, true> {
          *
          * @param rank the rank of the new group.
          */
-        AbelianGroup(unsigned rank);
+        AbelianGroup(unsigned long rank);
         /**
          * Creates a new group with the given rank and invariant factors.
          *
@@ -131,7 +131,7 @@ class AbelianGroup : public ShortOutput<AbelianGroup, true> {
          * is greater than 1 and divides the invariant factor after it.
          */
         template <typename T>
-        AbelianGroup(unsigned rank, std::initializer_list<T> invFac);
+        AbelianGroup(unsigned long rank, std::initializer_list<T> invFac);
         /**
          * Creates a new group with the given rank and invariant factors.
          *
@@ -152,7 +152,7 @@ class AbelianGroup : public ShortOutput<AbelianGroup, true> {
          * is greater than 1 and divides the invariant factor after it.
          */
         template <typename Container>
-        AbelianGroup(unsigned rank, const Container& invFac);
+        AbelianGroup(unsigned long rank, const Container& invFac);
         /**
          * Creates the abelian group defined by the given presentation matrix.
          *
@@ -231,7 +231,7 @@ class AbelianGroup : public ShortOutput<AbelianGroup, true> {
          *
          * @param extraRank the extra rank to add; this defaults to 1.
          */
-        void addRank(int extraRank = 1);
+        void addRank(long extraRank = 1);
         /**
          * Adds the given torsion element to the group.
          *
@@ -348,7 +348,7 @@ class AbelianGroup : public ShortOutput<AbelianGroup, true> {
          *
          * @return the number of included copies of \e Z.
          */
-        unsigned rank() const;
+        unsigned long rank() const;
         /**
          * Returns the rank in the group of the torsion term of given degree.
          * If the given degree is <i>d</i>, this routine will return the
@@ -365,7 +365,7 @@ class AbelianGroup : public ShortOutput<AbelianGroup, true> {
          * @param degree the degree of the torsion term to query.
          * @return the rank in the group of the given torsion term.
          */
-        unsigned torsionRank(const Integer& degree) const;
+        unsigned long torsionRank(const Integer& degree) const;
         /**
          * Returns the rank in the group of the torsion term of given degree.
          * If the given degree is <i>d</i>, this routine will return the
@@ -382,7 +382,7 @@ class AbelianGroup : public ShortOutput<AbelianGroup, true> {
          * @param degree the degree of the torsion term to query.
          * @return the rank in the group of the given torsion term.
          */
-        unsigned torsionRank(unsigned long degree) const;
+        unsigned long torsionRank(unsigned long degree) const;
         /**
          * Returns the number of invariant factors that describe the
          * torsion elements of this group.
@@ -425,7 +425,7 @@ class AbelianGroup : public ShortOutput<AbelianGroup, true> {
          * @return \c true if and only if this is the free abelian group
          * of rank \a r.
          */
-        bool isFree(unsigned r) const;
+        bool isFree(unsigned long r) const;
         /**
          * Determines whether this is the non-trivial cyclic group on
          * the given number of elements.
@@ -563,11 +563,11 @@ void swap(AbelianGroup& lhs, AbelianGroup& rhs) noexcept;
 inline AbelianGroup::AbelianGroup() {
 }
 
-inline AbelianGroup::AbelianGroup(unsigned rank) : rank_(rank) {
+inline AbelianGroup::AbelianGroup(unsigned long rank) : rank_(rank) {
 }
 
 template <typename T>
-inline AbelianGroup::AbelianGroup(unsigned rank,
+inline AbelianGroup::AbelianGroup(unsigned long rank,
         std::initializer_list<T> invFac) : rank_(rank) {
     if (invFac.size() > 0) {
         auto it = std::rbegin(invFac);
@@ -587,7 +587,7 @@ inline AbelianGroup::AbelianGroup(unsigned rank,
 }
 
 template <typename Container>
-inline AbelianGroup::AbelianGroup(unsigned rank, const Container& invFac) :
+inline AbelianGroup::AbelianGroup(unsigned long rank, const Container& invFac) :
         rank_(rank) {
     if (! invFac.empty()) {
         auto it = invFac.rbegin();
@@ -611,7 +611,7 @@ inline void AbelianGroup::swap(AbelianGroup& other) noexcept {
     revInvFactors_.swap(other.revInvFactors_);
 }
 
-inline void AbelianGroup::addRank(int extraRank) {
+inline void AbelianGroup::addRank(long extraRank) {
     rank_ += extraRank;
 }
 
@@ -628,11 +628,11 @@ inline void AbelianGroup::addTorsionElement(unsigned long degree,
         addTorsion(d);
 }
 
-inline unsigned AbelianGroup::rank() const {
+inline unsigned long AbelianGroup::rank() const {
     return rank_;
 }
 
-inline unsigned AbelianGroup::torsionRank(unsigned long degree) const {
+inline unsigned long AbelianGroup::torsionRank(unsigned long degree) const {
     return torsionRank(Integer(degree));
 }
 
@@ -652,7 +652,7 @@ inline bool AbelianGroup::isZ() const {
     return (rank_ == 1 && revInvFactors_.empty());
 }
 
-inline bool AbelianGroup::isFree(unsigned r) const {
+inline bool AbelianGroup::isFree(unsigned long r) const {
     return (rank_ == r && revInvFactors_.empty());
 }
